% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{average_colors}
\alias{average_colors}
\title{Determine the numerical average of a vector of colors}
\usage{
average_colors(x, method="RGB")
}
\arguments{
\item{x}{vector of color names to average}

\item{method}{character, specificying the color space in which to average. Currently only "RGB" is accepted.}
}
\value{
character, the hexadecimal representation of the color
}
\description{
Convert a vector of colors into a single color that averages their numerical values of RGB (red-green-blue) or Lab (lightness-a-b) characteristics.
}
\details{
Solution derived from answer by stackoverflow user Deleplace at https://stackoverflow.com/questions/14482226/how-can-i-get-the-color-halfway-between-two-colors. Implementation reproduced from [BenaroyaResearch](https://github.com/BenaroyaResearch/miscHelpers/).
}
