% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicFrame.R
\name{DynamicFrame}
\alias{DynamicFrame}
\title{Dynamic expanding data frame class}
\description{
Dynamic expanding data frame class

Dynamic expanding data frame class
}
\details{
Designed to double space in memory when needed instead of always expanding the list.
Dynamically assembles the resulting data frame from a list of elements when called.

This is good if we:
1. Don't know the total data size ahead of time
2. Add rows incrementally and often, e.g., in a loop
3. Retrieve the full data infrequently, e.g., while plotting.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{df}}{Retrieve data as a tibble}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DynamicFrame-new}{\code{DynamicFrame$new()}}
\item \href{#method-DynamicFrame-append}{\code{DynamicFrame$append()}}
\item \href{#method-DynamicFrame-clone}{\code{DynamicFrame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicFrame-new"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicFrame-new}{}}}
\subsection{Method \code{new()}}{
Initialise a data frame with some data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicFrame$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Add arguments as rows of a dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicFrame-append"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicFrame-append}{}}}
\subsection{Method \code{append()}}{
Add rows to the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicFrame$append(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Data as if passed to `tibble::tibble(...)`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
