% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoMalaria-package.R
\docType{package}
\name{GeoMalaria-package}
\alias{GeoMalaria}
\alias{GeoMalaria-package}
\title{GeoMalaria: Spatial Malaria Simulation}
\description{
Spatial malaria simulation with individual humans and probabilistic mosquitoes.
}
\keyword{internal}
