% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{Human}
\alias{Human}
\title{Shell class to hold human attributes and methods}
\description{
Shell class to hold human attributes and methods

Shell class to hold human attributes and methods
}
\details{
To use a different mosquito, either:
1. Use `$new(...)$` to initialise an instance of this class with different constants
2. Use the `$set(...)` method to overwrite class methods, then initialise an instance
3. Create a child class


Calculates the blood parasite density after infection in counts per microlitre.
Specify either `population` AND `t`, OR `dt`



Calculates a vector of probability of infecting mosquito with gametocytes in the grid cell.
Specify either `population` AND `t`, OR `dt`



Full immunity then decays to 50% 
scales # bites not # susceptible people. E.g. no one retains complete immunity
after decay, everyone retains partial protection.

Specify either `population` AND `t`, OR `dt`
}
\examples{

## ------------------------------------------------
## Method `Human$plot`
## ------------------------------------------------

human = Human$new()
plot(human)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{duration_infectivity}}{Constrain infectious period}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Human-new}{\code{Human$new()}}
\item \href{#method-Human-parasitaemia}{\code{Human$parasitaemia()}}
\item \href{#method-Human-has_detection}{\code{Human$has_detection()}}
\item \href{#method-Human-has_fever}{\code{Human$has_fever()}}
\item \href{#method-Human-has_death}{\code{Human$has_death()}}
\item \href{#method-Human-infectivity}{\code{Human$infectivity()}}
\item \href{#method-Human-immunity}{\code{Human$immunity()}}
\item \href{#method-Human-plot}{\code{Human$plot()}}
\item \href{#method-Human-clone}{\code{Human$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-new"></a>}}
\if{latex}{\out{\hypertarget{method-Human-new}{}}}
\subsection{Method \code{new()}}{
Initialise a mosquito class, optionally overwriting the default parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$new(duration_infectivity = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments of parameter constants or functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-parasitaemia"></a>}}
\if{latex}{\out{\hypertarget{method-Human-parasitaemia}{}}}
\subsection{Method \code{parasitaemia()}}{
Lifecycle stages for P. vivax in population after inoculation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$parasitaemia(population = NULL, t = NULL, dt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{Dataframe of population with a `t_infection` column}

\item{\code{t}}{Simulation time}

\item{\code{dt}}{Time since infection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-has_detection"></a>}}
\if{latex}{\out{\hypertarget{method-Human-has_detection}{}}}
\subsection{Method \code{has_detection()}}{
Criteria to be laboratory-detectable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$has_detection(d_parasitaemia)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-has_fever"></a>}}
\if{latex}{\out{\hypertarget{method-Human-has_fever}{}}}
\subsection{Method \code{has_fever()}}{
Criteria to show symptoms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$has_fever(d_parasitaemia)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-has_death"></a>}}
\if{latex}{\out{\hypertarget{method-Human-has_death}{}}}
\subsection{Method \code{has_death()}}{
Criteria for death
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$has_death(d_parasitaemia)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-infectivity"></a>}}
\if{latex}{\out{\hypertarget{method-Human-infectivity}{}}}
\subsection{Method \code{infectivity()}}{
Lifecycle stages for P. vivax in population after inoculation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$infectivity(population = NULL, t = NULL, dt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{Dataframe of population with a `t_infection` column}

\item{\code{t}}{Simulation time}

\item{\code{dt}}{Time since infection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-immunity"></a>}}
\if{latex}{\out{\hypertarget{method-Human-immunity}{}}}
\subsection{Method \code{immunity()}}{
Human immunity profile over time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$immunity(population = NULL, t = NULL, dt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{Dataframe of population with a `t_infection` column}

\item{\code{t}}{Simulation time}

\item{\code{dt}}{Time since infection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Human-plot}{}}}
\subsection{Method \code{plot()}}{
Plot characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$plot()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{human = Human$new()
plot(human)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Human-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Human-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
