% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{label_auto}
\alias{label_auto}
\title{Format numbers with tidy k/M/B/T/pct suffixes for ggplot axes}
\usage{
label_auto(x)
}
\arguments{
\item{x}{Numeric vector, to turn into labels}
}
\description{
Format numbers with tidy k/M/B/T/pct suffixes for ggplot axes
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=cyl, y=mpg)) +
  scale_y_continuous(labels = label_auto)
}
