% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito.R
\name{Mosquito}
\alias{Mosquito}
\title{Shell class to hold species attributes and methods}
\description{
Shell class to hold species attributes and methods

Shell class to hold species attributes and methods
}
\details{
To use a different mosquito, either:
1. Use `$new(...)$` to initialise an instance of this class with different constants
2. Use the `$set(...)` method to overwrite class methods, then initialise an instance
3. Create a child class


Mosquito density calculated if they started at a point and traveled `dx` and
`dy` in time `dt`. Variance is constant at (km/day)^2. Uses the Laplace isometric
diffusion solution.
}
\examples{

## ------------------------------------------------
## Method `Mosquito$plot`
## ------------------------------------------------

mos = Mosquito$new()
plot(mos)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{distribution}}{Raster of spatial mosquito counts}

\item{\code{death_rate}}{Proportion of mosquitoes that die per day, simulated continuously}

\item{\code{travel}}{Constant of mosquito travel in variance of location after one day}

\item{\code{bite_rate}}{Bites per mosquito per density}

\item{\code{max_lifespan}}{Automatic: ignore mosquitoes after this duration after biting}

\item{\code{max_flight_range}}{Automatic: ignore mosquitoes this distance from original bite site}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mosquito-new}{\code{Mosquito$new()}}
\item \href{#method-Mosquito-survival}{\code{Mosquito$survival()}}
\item \href{#method-Mosquito-sporogony}{\code{Mosquito$sporogony()}}
\item \href{#method-Mosquito-migration}{\code{Mosquito$migration()}}
\item \href{#method-Mosquito-plot}{\code{Mosquito$plot()}}
\item \href{#method-Mosquito-clone}{\code{Mosquito$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mosquito-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mosquito-new}{}}}
\subsection{Method \code{new()}}{
Initialise a mosquito class, optionally overwriting the default parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito$new(distribution, death_rate = 0.25, travel = 1^2, bite_rate = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments of parameter constants or functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mosquito-survival"></a>}}
\if{latex}{\out{\hypertarget{method-Mosquito-survival}{}}}
\subsection{Method \code{survival()}}{
Proportion of mosquitoes that survive over time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito$survival(dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dt}}{Time since mosquito inoculation in days}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mosquito-sporogony"></a>}}
\if{latex}{\out{\hypertarget{method-Mosquito-sporogony}{}}}
\subsection{Method \code{sporogony()}}{
Proportion of mosquitoes with sporozoites over time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito$sporogony(dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dt}}{Time since mosquito inoculation in days}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mosquito-migration"></a>}}
\if{latex}{\out{\hypertarget{method-Mosquito-migration}{}}}
\subsection{Method \code{migration()}}{
Mosquito density
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito$migration(dx, dy, dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dx}}{Distance traveled horizontally in km}

\item{\code{dy}}{Distance traveled vertically}

\item{\code{dt}}{Time since origin in days}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mosquito-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Mosquito-plot}{}}}
\subsection{Method \code{plot()}}{
Print mosquito summary


Plot mosquito characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito$plot()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mos = Mosquito$new()
plot(mos)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mosquito-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Mosquito-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
