% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_extract}
\alias{my_extract}
\title{Extract values from Raster objects}
\usage{
my_extract(...)
}
\arguments{
\item{...}{Arguments to be passed to `raster::extract`}
}
\description{
See `raster::extract` but returns NA if there are no values to be extracted.
}
