% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.Simulation}
\alias{plot.Simulation}
\title{Plot current state of a simulation}
\usage{
plot.Simulation(x, ...)
}
\arguments{
\item{x}{Simulation object}

\item{...}{Additional arguments}
}
\description{
Plots map with epicurve
}
