% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_epicurve}
\alias{plot_epicurve}
\title{Plot epidemic curve}
\usage{
plot_epicurve(sim, t = NULL, ...)
}
\arguments{
\item{sim}{Simulation object}

\item{t}{Optional, plot epicurve as it was at this time}

\item{...}{Unused}
}
\description{
Plot epidemic curve
}
