% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{plot_human_clinical_outcomes}
\alias{plot_human_clinical_outcomes}
\title{Plot human clinical outcomes}
\usage{
plot_human_clinical_outcomes(human)
}
\arguments{
\item{human}{Human object}
}
\description{
Plot human clinical outcomes
}
