% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{plot_human_infectivity}
\alias{plot_human_infectivity}
\title{Plot human infectivity curve}
\usage{
plot_human_infectivity(human)
}
\arguments{
\item{human}{Human object}
}
\description{
Plot human infectivity curve
}
