% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_init}
\alias{plot_init}
\title{Plot initial state of a simulation}
\usage{
plot_init(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\description{
Valid at any time period, output should not depend on whether simulation is run
}
