% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito.R
\name{plot_mosquito_infectivity}
\alias{plot_mosquito_infectivity}
\title{Plot mosquito infectivity curve}
\usage{
plot_mosquito_infectivity(mos)
}
\arguments{
\item{mos}{Mosquito object}
}
\description{
Plot mosquito infectivity curve
}
\examples{
mos = Mosquito$new()
plot_mosquito_infectivity(mos)
}
