% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito.R
\name{plot_mosquito_migration}
\alias{plot_mosquito_migration}
\title{Plot spread of mosquitoes over distance}
\usage{
plot_mosquito_migration(mos)
}
\arguments{
\item{mos}{Mosquito object}
}
\description{
Plot spread of mosquitoes over distance
}
\examples{
mos = Mosquito$new()
plot_mosquito_migration(mos)
}
