% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_state}
\alias{plot_state}
\title{Plot map of simulation}
\usage{
plot_state(sim, t = NULL, background = "EIR", ...)
}
\arguments{
\item{sim}{Simulation object}

\item{t}{Optional, plot simulation as it was at this time}

\item{background}{Either 'EIR' or 'mosquito'}

\item{...}{Unused}
}
\description{
Plot map of simulation
}
