% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{print.Simulation}
\alias{print.Simulation}
\title{Print simulation}
\usage{
\method{print}{Simulation}(x, ...)
}
\arguments{
\item{x}{Simulation object}

\item{...}{Additional arguments}
}
\description{
Print simulation
}
