% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_anim}
\alias{plot_anim}
\title{Animates a simulation}
\usage{
plot_anim(sim, t = NULL, file = NULL, ...)
}
\arguments{
\item{sim}{Simulation object}

\item{t}{Optional, vector of times to plot}

\item{file}{Optional, write video to a destination}

\item{...}{Additional arguments}
}
\description{
Plots map with epicurve. Pre-calculates ranges to ensure frames are consistent.
}
