% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_mosquito_infectivity}
\alias{plot_mosquito_infectivity}
\title{Plot mosquito infectivity curve}
\usage{
plot_mosquito_infectivity(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\description{
Plot mosquito infectivity curve
}
