% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_mosquito_migration}
\alias{plot_mosquito_migration}
\title{Plot spread of mosquitoes over distance}
\usage{
plot_mosquito_migration(sim)
}
\arguments{
\item{sim}{Simulation object}
}
\description{
Plot spread of mosquitoes over distance
}
