% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Initialize a simulation instance}
\description{
Initialize a simulation instance

Initialize a simulation instance
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulation-new}{\code{Simulation$new()}}
\item \href{#method-Simulation-iterate}{\code{Simulation$iterate()}}
\item \href{#method-Simulation-set_mosquito_raster}{\code{Simulation$set_mosquito_raster()}}
\item \href{#method-Simulation-calculate_EIR}{\code{Simulation$calculate_EIR()}}
\item \href{#method-Simulation-mosquito_survival}{\code{Simulation$mosquito_survival()}}
\item \href{#method-Simulation-mosquito_sporogony}{\code{Simulation$mosquito_sporogony()}}
\item \href{#method-Simulation-mosquito_migration}{\code{Simulation$mosquito_migration()}}
\item \href{#method-Simulation-human_infectivity}{\code{Simulation$human_infectivity()}}
\item \href{#method-Simulation-human_immunity}{\code{Simulation$human_immunity()}}
\item \href{#method-Simulation-human_schedule_relapse}{\code{Simulation$human_schedule_relapse()}}
\item \href{#method-Simulation-clone}{\code{Simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(
  humans,
  locations,
  mosquito_raster,
  duration_human_infectivity,
  mosquito_death_rate,
  bite_rate,
  p_relapse = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{humans}}{dataframe of residents in the area}

\item{\code{mosquito_raster}}{raster of the number of mosquitoes per cell}

\item{\code{duration_human_infectivity}}{maximum infectious period of a human in days}

\item{\code{mosquito_death_rate}}{proportion of mosquitoes that die per day}

\item{\code{bite_rate}}{probability of each mosquito biting a human in a cell
Recompute location mosquito densities from a raster}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-iterate}{}}}
\subsection{Method \code{iterate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$iterate(dt, debug = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-set_mosquito_raster"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-set_mosquito_raster}{}}}
\subsection{Method \code{set_mosquito_raster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$set_mosquito_raster(mosquito_raster)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
mosquito_raster for further assignment
Calculate EIR at a location with X and Y elements
Proportion of mosquitoes that survive over time
Proportion of mosquitoes with sporozoites over time
Lifecycle stages for P. vivax in humans after inoculation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-calculate_EIR"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-calculate_EIR}{}}}
\subsection{Method \code{calculate_EIR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$calculate_EIR(loc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-mosquito_survival"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-mosquito_survival}{}}}
\subsection{Method \code{mosquito_survival()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$mosquito_survival(dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-mosquito_sporogony"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-mosquito_sporogony}{}}}
\subsection{Method \code{mosquito_sporogony()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$mosquito_sporogony(dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-mosquito_migration"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-mosquito_migration}{}}}
\subsection{Method \code{mosquito_migration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$mosquito_migration(dx, dy, dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-human_infectivity"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-human_infectivity}{}}}
\subsection{Method \code{human_infectivity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$human_infectivity(dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{days since inoculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of probability of infecting mosquito with gametocytes in the grid cell
Human immunity profile over time

Full immunity then decays to 50% 
scales # bites not # susceptible people. E.g. no one retains complete immunity
after decay, everyone retains partial protection
Human relapse distribution

Schedule a relapse infection for number of days after the primary infection.
NA value is no relapse.
shape = (mean / sd)^2
rate =  mean / sd^2
Representation of humans with only one set of coordinates

Just places people at their highest proportion location
Represent humans at all of their coordinates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-human_immunity"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-human_immunity}{}}}
\subsection{Method \code{human_immunity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$human_immunity(dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-human_schedule_relapse"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-human_schedule_relapse}{}}}
\subsection{Method \code{human_schedule_relapse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$human_schedule_relapse(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
