% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Malaria simulation class}
\description{
Malaria simulation class

Malaria simulation class
}
\details{
Agent-based simulation for P. vivax malaria in a spatial context


Advance simulation by `dt`. Calculates the current state of population, infects
mosquitoes and population simultaneously, updates population and mosquitoes, then
stores observations for analysis.


Updates simulation's raster and densities at each location


Calculate EIR at a location with x and y elements



Schedule a relapse infection for number of days after the primary infection.
Returns `Inf` value if no relapse.
shape = (mean / sd)^2
rate =  mean / sd^2



Schedule a recovery for number of days after the primary infection. Can
also be used to schedule the success of treatment.
Returns `Inf` value if no complete clearance



Plots map with epicurve or the starting parameters if simulation is in its initialised state


Just places people at their highest proportion location
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{t}}{Current simulation time in days}

\item{\code{min_dt}}{Smallest time step used so far}

\item{\code{mosquito_raster}}{Raster object of mosquito counts}

\item{\code{sporozoite_infection_rate}}{Proportion of infectious mosquito bites that can result in infection}

\item{\code{p_relapse}}{Probability of an infection scheduling a relapse}

\item{\code{mean_recovery}}{Recovery parameter mean=1/rate. Could be changed for other distributions or functions}

\item{\code{vis_raster}}{Raster to visualise any spatial fields on}

\item{\code{population}}{Data frame of population}

\item{\code{locations}}{Data frame of human locations}

\item{\code{mosquito_infections}}{Data frame of current mosquito infection events}

\item{\code{relapse_shape}}{Shape constant of the human relapse distribution}

\item{\code{relapse_rate}}{Rate constant of the human relapse distribution}

\item{\code{log}}{Character vector of logging options. Options include "linelist", "compartment", and/or "EIR"}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{population_collapse}}{Represent population with only one set of coordinates}

\item{\code{population_expand}}{Represent population at all of their coordinates}

\item{\code{linelist}}{Safely access linelist if logged}

\item{\code{compartment}}{Safely access compartments if logged}

\item{\code{EIR}}{Entomological inoculation rate by coordinates over time}

\item{\code{vis_grid}}{Data frame version of `vis_raster`}

\item{\code{attack_rate}}{Proportion of population who have been infected at least once}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulation-new}{\code{Simulation$new()}}
\item \href{#method-Simulation-iterate}{\code{Simulation$iterate()}}
\item \href{#method-Simulation-set_mosquito_raster}{\code{Simulation$set_mosquito_raster()}}
\item \href{#method-Simulation-calculate_EIR}{\code{Simulation$calculate_EIR()}}
\item \href{#method-Simulation-schedule_relapse}{\code{Simulation$schedule_relapse()}}
\item \href{#method-Simulation-schedule_recovery}{\code{Simulation$schedule_recovery()}}
\item \href{#method-Simulation-plot}{\code{Simulation$plot()}}
\item \href{#method-Simulation-print}{\code{Simulation$print()}}
\item \href{#method-Simulation-clone}{\code{Simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a simulation instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(
  population,
  locations,
  mosquito = Mosquito$new(),
  mosquito_raster,
  human = Human$new(),
  p_relapse = 0,
  mean_recovery = 14,
  log_options = c("linelist", "compartment")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{dataframe of residents in the area}

\item{\code{locations}}{dataframe of human locations}

\item{\code{mosquito}}{class for mosquitoes}

\item{\code{mosquito}}{class for mosquitoes}

\item{\code{mosquito_raster}}{raster of the number of mosquitoes per cell}

\item{\code{human}}{class for humans}

\item{\code{human}}{class for humans}

\item{\code{p_relapse}}{Proportion of infections that trigger relapses}

\item{\code{mean_recovery}}{Unused}

\item{\code{log_options}}{Logging options}

\item{\code{mosquito_death_rate}}{proportion of mosquitoes that die per day}

\item{\code{bite_rate}}{probability of each mosquito biting a human in a cell}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-iterate}{}}}
\subsection{Method \code{iterate()}}{
Iterate simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$iterate(dt, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dt}}{Time to advance simulation by in days}

\item{\code{debug}}{Optional, print debug information}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-set_mosquito_raster"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-set_mosquito_raster}{}}}
\subsection{Method \code{set_mosquito_raster()}}{
Recompute location mosquito densities from a raster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$set_mosquito_raster(mosquito_raster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mosquito_raster}}{Raster of mosquito densities}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-calculate_EIR"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-calculate_EIR}{}}}
\subsection{Method \code{calculate_EIR()}}{
Entomological inoculation rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$calculate_EIR(loc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loc}}{List or dataframe with `x` and `y` columns}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numeric vector of EIR
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-schedule_relapse"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-schedule_relapse}{}}}
\subsection{Method \code{schedule_relapse()}}{
Human relapse distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$schedule_relapse(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of people to schedule relapses for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-schedule_recovery"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-schedule_recovery}{}}}
\subsection{Method \code{schedule_recovery()}}{
Human recovery distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$schedule_recovery(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of people to schedule relapses for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-plot}{}}}
\subsection{Method \code{plot()}}{
Plot current state of a simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$plot(t = NULL, init = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Optional, plot simulation as it was at this time}

\item{\code{init}}{Logical, show initialisation state with mosquito raster}

\item{\code{...}}{Additional arguments}

\item{\code{x}}{Simulation object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-print"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-print}{}}}
\subsection{Method \code{print()}}{
Print simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments}

\item{\code{x}}{Simulation object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
